New BasePage created. Added funcionality for switching context on mobile
